feat: update cf facade example
