+++ packages/sheets-conditional-formatting/src/facade/index.ts
@@ -17,8 +17,9 @@
 import './f-range';
 import './f-workbook';
 import './f-worksheet';
+import './f-enum';
 
-export { FConditionalFormattingBuilder } from './conditional-formatting-builder';
+export { FConditionalFormattingBuilder } from './f-conditional-formatting-builder';
 export type { IFRangeConditionalFormattingMixin } from './f-range';
 export type { FWorkbookConditionalFormattingMixin } from './f-workbook';
